+++ backend/app/__init__.py
@@ -1,4 +1,4 @@
 """Mosaico Backend Application"""
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 __app_name__ = "mosaico"